third_party/blink/web_tests/external/wpt/css/css-box/margin-trim/computed-margin-values/flexbox-column-multi-line-block-expected.txt

This is a testharness.js-based test.
[FAIL] flexbox > item 1
  assert_equals: \n<item data-offset-y="8" data-expected-margin-top="0" data-expected-margin-bottom="10"></item>\noffsetTop expected 8 but got 18
[FAIL] flexbox > item 2
  assert_equals: \n<item data-offset-y="78" data-expected-margin-top="10" data-expected-margin-bottom="0"></item>\noffsetTop expected 78 but got 18
[FAIL] flexbox > item 3
  assert_equals: \n<item data-offset-y="8" data-expected-margin-top="0" data-expected-margin-bottom="10"></item>\noffsetTop expected 8 but got 18
[FAIL] flexbox > item 4
  assert_equals: \n<item data-offset-y="78" data-expected-margin-top="10" data-expected-margin-bottom="0"></item>\noffsetTop expected 78 but got 18
Harness: the test ran to completion.
